Split up proto files
