+++ adapters/handlers/grpc/batch_parse_request_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/weaviate/weaviate/entities/models"
 	"github.com/weaviate/weaviate/entities/schema"
-	"github.com/weaviate/weaviate/grpc"
+	"github.com/weaviate/weaviate/grpc/generated/grpc"
 )
 
 const (
