+++ grpc/base.proto
@@ -1,0 +1,31 @@
+syntax = "proto3";
+
+package base;
+option go_package = "github.com/weaviate/weaviate/grpc/generated/grpc";
+
+enum ConsistencyLevel {
+  CONSISTENCY_LEVEL_UNSPECIFIED = 0;
+  CONSISTENCY_LEVEL_ONE = 1;
+  CONSISTENCY_LEVEL_QUORUM = 2;
+  CONSISTENCY_LEVEL_ALL = 3;
+}
+
+message NumberArrayProperties {
+  repeated double values = 1;
+  string prop_name = 2;
+}
+
+message IntArrayProperties {
+  repeated int64 values = 1;
+  string prop_name = 2;
+}
+
+message TextArrayProperties {
+  repeated string values = 1;
+  string prop_name = 2;
+}
+
+message BooleanArrayProperties {
+  repeated bool values = 1;
+  string prop_name = 2;
+}
