+++ grpc/search_get.proto
@@ -1,0 +1,259 @@
+syntax = "proto3";
+
+package search_get;
+
+import "google/protobuf/struct.proto";
+import "grpc/base.proto";
+option go_package = "github.com/weaviate/weaviate/grpc/generated/grpc";
+
+
+
+message SearchRequest {
+  string class_name = 1;
+  uint32 limit = 2;
+  AdditionalProperties additional_properties = 3;
+  NearVectorParams near_vector = 4;
+  NearObjectParams near_object = 5;
+  Properties properties = 6;
+  HybridSearchParams hybrid_search =7;
+  BM25SearchParams bm25_search =8;
+  uint32 offset = 9;
+  uint32 autocut = 10;
+  string after = 11;
+  string tenant = 12;
+  optional Filters filters = 13;
+  optional NearTextSearchParams near_text = 14;
+  optional NearImageSearchParams near_image = 15;
+  optional NearAudioSearchParams near_audio = 16;
+  optional NearVideoSearchParams near_video = 17;
+  optional base.ConsistencyLevel consistency_level = 18;
+  optional GenerativeSearch generative = 19;
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated SortBy sort_by = 20;
+  optional GroupBy group_by = 21;
+}
+
+message GroupBy {
+  // currently only supports one entry (eg just properties, no refs). But might
+  // be extended in the future.
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated string path = 1;
+  int32 number_of_groups = 2;
+  int32 objects_per_group = 3;
+}
+
+message SortBy {
+  bool ascending = 1;
+  // currently only supports one entry (eg just properties, no refs). But the
+  // weaviate datastructure already has paths in it and this makes it easily
+  // extendable in the future
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated string path = 2;
+}
+
+message GenerativeSearch {
+  string single_response_prompt = 1;
+  string grouped_response_task = 2;
+  repeated string grouped_properties = 3;
+}
+
+message TextArray {
+  repeated string values = 1;
+}
+message IntArray {
+  repeated int64 values = 1;
+}
+message NumberArray {
+  repeated double values = 1;
+}
+message BooleanArray {
+  repeated bool values = 1;
+}
+
+message Filters {
+  enum Operator {
+    OPERATOR_UNSPECIFIED = 0;
+    OPERATOR_EQUAL = 1;
+    OPERATOR_NOT_EQUAL = 2;
+    OPERATOR_GREATER_THAN = 3;
+    OPERATOR_GREATER_THAN_EQUAL = 4;
+    OPERATOR_LESS_THAN = 5;
+    OPERATOR_LESS_THAN_EQUAL = 6;
+    OPERATOR_AND = 7;
+    OPERATOR_OR = 8;
+    OPERATOR_WITHIN_GEO_RANGE = 9;
+    OPERATOR_LIKE = 10;
+    OPERATOR_IS_NULL = 11;
+    OPERATOR_CONTAINS_ANY = 12;
+    OPERATOR_CONTAINS_ALL = 13;
+  }
+
+
+  Operator operator = 1;
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated string on = 2;
+  repeated Filters filters = 3;
+  oneof test_value{
+    string value_text = 4;
+    int64 value_int = 5;
+    bool value_boolean = 6;
+    float value_number = 7;
+    TextArray value_text_array = 9;
+    IntArray value_int_array = 10;
+    BooleanArray value_boolean_array = 11;
+    NumberArray value_number_array = 12;
+  };
+}
+
+
+message AdditionalProperties {
+  bool uuid = 1;
+  bool vector = 2;
+  // protolint:disable:next FIELD_NAMES_LOWER_SNAKE_CASE
+  bool creationTimeUnix = 3;
+  // protolint:disable:next FIELD_NAMES_LOWER_SNAKE_CASE
+  bool lastUpdateTimeUnix = 4;
+  bool distance = 5;
+  bool certainty = 6;
+  bool score = 7;
+  // protolint:disable:next FIELD_NAMES_LOWER_SNAKE_CASE
+  bool explainScore = 8;
+  bool is_consistent = 9;
+}
+
+
+message Properties {
+  repeated string non_ref_properties = 1;
+  repeated RefProperties ref_properties = 2;
+}
+
+message HybridSearchParams {
+  string query = 1;
+  repeated string properties = 2;
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated float vector = 3;
+  float alpha = 4;
+  enum FusionType {
+    FUSION_TYPE_UNSPECIFIED = 0;
+    FUSION_TYPE_RANKED = 1;
+    FUSION_TYPE_RELATIVE_SCORE = 2;
+  }
+  FusionType fusion_type = 5;
+}
+
+message NearTextSearchParams {
+  message Move{
+    float force = 1;
+    repeated string concepts = 2;
+    repeated string uuids = 3;
+  }
+
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated string query = 1;
+  optional double certainty = 2;
+  optional double distance = 3;
+  optional Move move_to = 4;
+  optional Move move_away = 5;
+};
+
+message NearImageSearchParams {
+  string image = 1;
+  optional double certainty = 2;
+  optional double distance = 3;
+};
+
+message NearAudioSearchParams {
+  string audio = 1;
+  optional double certainty = 2;
+  optional double distance = 3;
+};
+
+message NearVideoSearchParams {
+  string video = 1;
+  optional double certainty = 2;
+  optional double distance = 3;
+};
+
+message BM25SearchParams {
+  string query = 1;
+  repeated string properties = 2;
+}
+
+
+message RefProperties {
+  string reference_property = 2;
+  Properties linked_properties = 3;
+  AdditionalProperties metadata = 4;
+  string which_collection = 5;
+}
+
+message NearVectorParams {
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated float vector = 1;
+  optional double certainty = 2;
+  optional double distance = 3;
+}
+
+message NearObjectParams {
+  string id = 1;
+  optional double certainty = 2;
+  optional double distance = 3;
+}
+
+message SearchReply {
+  repeated SearchResult results = 1;
+  float took = 2;
+  string generative_grouped_result = 3;
+  repeated GroupByResults group_by_results = 4;
+}
+
+message GroupByResults {
+  string name = 1;
+  float min_distance = 2;
+  float max_distance = 3;
+  int64 number_of_objects = 4;
+  repeated SearchResult objects = 5;
+}
+
+message SearchResult {
+  ResultProperties properties = 1;
+  ResultAdditionalProps additional_properties = 2;
+}
+
+message ResultAdditionalProps {
+  string id = 1;
+  // protolint:disable:next REPEATED_FIELD_NAMES_PLURALIZED
+  repeated float vector = 2;
+  int64 creation_time_unix = 3;
+  bool creation_time_unix_present = 4;
+  int64 last_update_time_unix = 5;
+  bool last_update_time_unix_present = 6;
+  float distance = 7;
+  bool distance_present = 8;
+  float certainty = 9;
+  bool certainty_present = 10;
+  float score = 11;
+  bool score_present = 12;
+  string explain_score = 13;
+  bool explain_score_present = 14;
+  optional bool is_consistent = 15;
+  string generative = 16;
+  bool generative_present = 17;
+}
+
+message ResultProperties {
+  google.protobuf.Struct non_ref_properties = 1;
+  repeated ReturnRefProperties ref_props = 2;
+  string class_name = 3;
+  ResultAdditionalProps metadata = 4;
+  repeated base.NumberArrayProperties number_array_properties = 5;
+  repeated base.IntArrayProperties int_array_properties = 6;
+  repeated base.TextArrayProperties text_array_properties = 7;
+  repeated base.BooleanArrayProperties boolean_array_properties = 8;
+}
+
+message ReturnRefProperties {
+  repeated ResultProperties properties = 1;
+  string prop_name = 2;
+}
+
